pyproject.toml
[project]
name = "fastapi-crud"
version = "0.1.0"
description = ""
authors = [
    {name = "leshakovalevskiy2002",email = "lesha_programmer02_gt@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
